+++ packages/sync-server/app/admin/programImporter/validation.js
@@ -1,10 +1,35 @@
-import { VITALS_DATA_ELEMENT_IDS, SURVEY_TYPES, CURRENTLY_AT_TYPES } from '@tamanu/constants';
-import { ValidationError } from '../errors';
-import { statkey, updateStat } from '../stats';
+import { Op } from 'sequelize';
+import { VITALS_DATA_ELEMENT_IDS, SURVEY_TYPES } from '@tamanu/constants';
+import { ImporterMetadataError, ValidationError } from '../errors';
 
 const REQUIRED_QUESTION_IDS = {
   [SURVEY_TYPES.VITALS]: Object.values(VITALS_DATA_ELEMENT_IDS),
 };
+
+async function ensureOnlyOneVitalsSurveyExists({ models }, surveyInfo) {
+  const vitalsCount = await models.Survey.count({
+    where: {
+      id: {
+        [Op.not]: surveyInfo.id,
+      },
+      survey_type: SURVEY_TYPES.VITALS,
+    },
+  });
+  if (vitalsCount > 0) {
+    throw new ImporterMetadataError('Only one vitals survey may exist at a time');
+  }
+}
+
+function ensureVitalsSurveyNonSensitive(surveyInfo) {
+  if (surveyInfo.isSensitive) {
+    throw new ImporterMetadataError('Vitals survey can not be sensitive');
+  }
+}
+
+export async function validateVitalsSurvey(context, surveyInfo) {
+  await ensureOnlyOneVitalsSurveyExists(context, surveyInfo);
+  ensureVitalsSurveyNonSensitive(surveyInfo);
+}
 
 export function ensureRequiredQuestionsPresent(surveyInfo, questionRecords) {
   const { surveyType, sheetName } = surveyInfo;
@@ -23,52 +48,5 @@ export function ensureRequiredQuestionsPresent(surveyInfo, questionRecords) {
       -2,
       `Survey missing required questions: ${missingQuestions.join(', ')}`,
     );
-  }
-}
-
-export async function validateCurrentlyAtQuestionsCorrectType(context, surveyInfo, questionRecords) {
-  const { programId, sheetName } = surveyInfo;
-
-  // TODO: why not Program.hasOne(ProgramRegistry)
-  const programRegistry = await context.models.ProgramRegistry.findOne({
-    where: {
-      programId,
-      // Shouldn't be able to accidentally hack around the check with a historical registry
-      // visibilityStatus: VISIBILITY_STATUSES.CURRENT,
-    },
-  });
-  if (!programRegistry) return;
-
-  const forbiddenFieldName =
-    programRegistry.currentlyAtType === CURRENTLY_AT_TYPES.VILLAGE
-      ? 'registrationCurrentlyAtFacility'
-      : 'registrationCurrentlyAtVillage';
-
-  const newErrors = questionRecords
-    .filter(({ model }) => model === 'SurveyScreenComponent')
-    .map(record => ({ ...record, config: JSON.parse(record.values.config) }))
-    // The correct format for the config will be validated later, just ignore them for now
-    .filter(config => config?.writeToPatient?.fieldName)
-    .filter(config => config.writeToPatient.fieldName === forbiddenFieldName)
-    .map(
-      ({ sheetRow }) =>
-        new ValidationError(
-          sheetName,
-          sheetRow,
-          `Error in config.fieldName: fieldName=${forbiddenFieldName} but program registry configured for ${programRegistry.currentlyAtType}`,
-        ),
-    );
-
-  // Don't throw an error here so that validation can continue
-  // Note this means that errored stats might double count
-  // these rows if they have another problem as well
-  if (newErrors.length > 0) {
-    updateStat(
-      context.stats,
-      statkey('SurveyScreenComponent', sheetName),
-      'errored',
-      newErrors.length,
-    );
-    context.errors.push(newErrors);
   }
 }
